Move dfsutil functions to dfs
